requirements.txt
flask==2.0.1
pymongo==4.1.1
python-dotenv==0.19.0
dnspython==2.1.0